Move MTP sequence-roll infrastructure into dedicated module
